Features and Bugfixes:
1) VAPP:  New Virtual Application type added: “Postmark” (filesystem benchmark)
2) VAPP:  Recovered full functionality for the Virtual Application  type “OpenDaytrader”, including
the use of the Rain Workload Toolkit for as the Load Generating process
3) VAPP: Now the determination of the JAVA_HOME environment variable is carried out through a new common function: set_java_home
3) CORE: Hardened the scripts for the building of Docker based Virtual Application images
4) ADAPTERS: Multiple bug fixes for the Kubernetes Adapter.
5) CORE: Improvements on the automatic plot generation R script, including a VM # vs Failure graph.

+++ scenarios/scripts/kub_cleanup.sh
@@ -10,7 +10,8 @@ TNSRC_ERROR=0
 function all_clear () {
 
     CB_CLOUD_NAME=$(echo "$CB_CLOUD_NAME" | tr '[:upper:]' '[:lower:]')
-    for NS in $(kubectl get namespaces | grep ${KUB_NAMESPACE_NAME} | awk '{ print $1 }')
+#    for NS in $(kubectl get namespaces | grep ${KUB_NAMESPACE_NAME} | awk '{ print $1 }')
+    for NS in $(kubectl get namespaces | awk '{ print $1 }')
     do
         for D in $(kubectl --namespace $NS get deployments | grep ${CB_CLOUD_NAME} | grep ${CB_USERNAME} | awk '{ print $1 }')
         do  
@@ -30,11 +31,16 @@ function all_clear () {
             kubectl --namespace $NS delete pods ${P}
         done        
 
-        echo "deleting namespace $NS"       
-        kubectl delete namespace $NS
+        echo $NS | grep ${KUB_NAMESPACE_NAME} 
+        
+        if [[ $? -eq 0 ]]
+        then
+            echo "deleting namespace $NS"       
+            kubectl delete namespace $NS
+        fi
     done
 
-    for Q in $(kubectl get quota | grep ${KUB_NAMESPACE_NAME} | awk '{ print $1 }')
+    for Q in $(kubectl get quota | grep ${KUB_QUOTA_NAME} | awk '{ print $1 }')
     do  
         echo "deleting quota $Q"
         kubectl delete quota ${Q}
